+++ code/system/sys.data.crdt/src/crdt.DocRef/DocRef.mts
@@ -1,4 +1,4 @@
-import { init } from './DocRef.init.mjs';
+import { init } from './DocRef.impl.mjs';
 import { toObject } from '../crdt.helpers';
 
 /**
