+++ code/system/sys.data.crdt/src/crdt.DocSync/DocSync.mts
@@ -1,4 +1,4 @@
-import { init } from './DocSync.init.mjs';
+import { init } from './DocSync.impl.mjs';
 import { toObject } from '../crdt.helpers';
 
 /**
